requirements.txt
Django==3.2.0
djangorestframework==3.13.1
jdatetime==3.7.0